backend/requirements.txt
fastapi
uvicorn
google-genai
pydantic
python-dotenv
mermaid-py
